requirements.txt
# FastAPI and Web Server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pgvector==0.2.3

# PDF Processing
PyMuPDF==1.23.8
python-magic==0.4.27

# AI and ML
langchain==0.0.350
openai>=1.0.0
numpy==1.26.2
scikit-learn==1.3.2

# Google Drive Integration
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
tqdm==4.66.1
pytest==7.4.3
httpx==0.25.2
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Optional: For development
ipython==8.18.1
jupyter==1.0.0
streamlit==1.28.2
plotly==5.18.0
pandas==2.1.3 